+++ urbansim_templates/io/tables.py
@@ -3,7 +3,9 @@ from __future__ import print_function
 try:
     import pathlib  # Python 3.4+
 except:
-    import os
+    pass
+
+import os
 
 import orca
 import pandas as pd
@@ -14,13 +16,12 @@ from urbansim_templates import modelmanager, __version__
 @modelmanager.template
 class TableFromDisk():
     """
-    Class for registering data tables. In the initial implementation, data can come from  
-    local CSV or HDF5 files.
+    Class for registering data tables from local CSV or HDF5 files.
     
-    An instance of this Table() template stores *instructions for loading a data table*, 
-    which can be saved as a yaml file. Running these instructions registers the table 
-    with Orca. Saved tables will be registered automatically when you initialize 
-    ModelManager, replacing the `datasources.py` scripts used in previous versions of 
+    An instance of this template class stores *instructions for loading a data table*, 
+    packaged into an Orca step. Running the instructions registers the table with Orca. 
+    Saved table registration steps will be run automatically when you initialize 
+    ModelManager, replacing the ``datasources.py`` scripts used in previous versions of 
     UrbanSim. 
     
     Tables should include a unique index, or a set of columns that jointly represent a 
@@ -30,24 +31,7 @@ class TableFromDisk():
     be able to use it as a join key. Following these naming conventions eliminates the 
     need for Orca "broadcasts".
     
-    Usage
-    -----
-    Create an empty class instance: `t = TableFromDisk()`.
-    
-    Give it some properties: `t.name = 'buildings'` etc. (These can also be passed when 
-    you create the object.)
-    
-    Register with ModelManager: `modelmanager.register(t)`. This registers the table 
-    loading instructions, runs them, and saves them to disk. They'll automatically be run 
-    the next time you initialize ModelManager.
-    
-    You can use all the standard ModelManager commands to get copies of the saved table 
-    loading instructions, modify them, and delete them:
-    
-    - `modelmanager.list_steps()`
-    - `t2 = modelmanager.get_step('name')`
-    - `modelmanager.register(t2)`
-    - `modelmanager.remove_step('name')`
+    All the parameters can also be set as properties of the class instance.
     
     Parameters
     ----------
@@ -101,11 +85,6 @@ class TableFromDisk():
     
     autorun : bool, default True
         Automatically run the step whenever it's registered with ModelManager.
-    
-    Properties and attributes
-    -------------------------
-    All the parameters listed above can also be get and set as properties of the class 
-    instance.
     
     """
     def __init__(self, 
@@ -205,15 +184,14 @@ class TableFromDisk():
             except:
                 value = os.path.normpath(value)
         self.__path = value
-        print(value)
             
 
     def run(self):
         """
         Register a data table with Orca.
         
-        Requires values to be set for 'source_type', 'name', and 'path'. CSV data also 
-        requires 'csv_index_cols'. 
+        Requires values to be set for ``source_type``, ``name``, and ``path``. CSV data 
+        also requires ``csv_index_cols``. 
         
         Returns
         -------