+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # UrbanSim Templates change log
 
+### 0.2.dev0 (2019-02-19)
+
+- adds first data i/o template: `urbansim_templates.io.TableFromDisk()`
+- adds support for `autorun` template property
+
 ### 0.1.1 (2019-02-05)
 
 - production release